Start with interior minimap graphics
Right now it's just a full copy of the map, with some specific areas
already marked. The list of the interiors to go through is on
interior-checklist.txt.

Source graphic files are going to be made in chunks of 4 since it's
easier to work like that, as many interiors bleed between two sectors.